First pass at Squiggle Viewer refactor
