+++ packages/components/src/components/SquiggleChart.tsx
@@ -1,6 +1,6 @@
 import { FC, memo } from "react";
 
-import { SqValuePath } from "@quri/squiggle-lang";
+import { SqValue, SqValuePath } from "@quri/squiggle-lang";
 import { RefreshIcon } from "@quri/ui";
 
 import { SquiggleViewer } from "../index.js";
@@ -10,15 +10,13 @@ import {
   StandaloneExecutionProps,
   useSquiggle,
 } from "../lib/hooks/useSquiggle.js";
-import { MessageAlert } from "./Alert.js";
 import { PartialPlaygroundSettings } from "./PlaygroundSettings.js";
-import { SquiggleErrorAlert } from "./SquiggleErrorAlert.js";
-import { SquiggleOutputViewer } from "./SquiggleOutputViewer/index.js";
-import { useGetSubvalueByPath } from "./SquiggleViewer/utils.js";
+import { ViewerMenuBar } from "./ViewerMenuBar/index.js";
+import { ViewerBody } from "./ViewerMenuBar/ViewerBody.js";
 
 export type SquiggleChartProps = {
   code: string;
-  rootPathOverride?: SqValuePath; // Note: This should be static. We don't support rootPathOverride to change once set.
+  rootPathOverride?: SqValuePath; // Note: This should be static. We don't support rootPathOverride to change once set. Used for Export pages on Squiggle Hub.
 } & (StandaloneExecutionProps | ProjectExecutionProps) &
   // `environment` is passed through StandaloneExecutionProps; this way we guarantee that it's not compatible with `project` prop
   Omit<PartialPlaygroundSettings, "environment">;
@@ -44,35 +42,40 @@ export const SquiggleChart: FC<SquiggleChartProps> = memo(
     });
 
     // TODO - if `<ViewerProvider>` is not set up (which is very possible) then calculator paths won't be resolved.
-    const getSubvalueByPath = useGetSubvalueByPath();
 
     if (!squiggleOutput) {
       return <RefreshIcon className="animate-spin" />;
     }
 
     if (rootPathOverride) {
-      const { output } = squiggleOutput;
-      if (!output.ok) {
-        return <SquiggleErrorAlert error={output.value} />;
-      }
-      const rootValue =
-        rootPathOverride.root === "result"
-          ? output.value.result
-          : output.value.bindings.asValue();
+      // TODO: It's awkward how this toggle both changes the UI and the data flow. Maybe we should have a separate component for this?
 
-      const value = getSubvalueByPath(rootValue, rootPathOverride);
-      if (value) {
-        return <SquiggleViewer value={value} />;
-      } else {
-        return <MessageAlert heading="Value is not defined" />;
-      }
+      return (
+        <ViewerBody
+          output={squiggleOutput.output}
+          mode={{ tag: "CustomResultPath", value: rootPathOverride }}
+          isRunning={isRunning}
+          viewer={(output) => (
+            <SquiggleViewer
+              value={output}
+              environment={environment}
+              {...settings}
+            />
+          )}
+        />
+      );
     } else {
       return (
-        <SquiggleOutputViewer
+        <ViewerMenuBar
           squiggleOutput={squiggleOutput}
           isRunning={isRunning}
-          environment={environment}
-          {...settings}
+          viewer={(output: SqValue) => (
+            <SquiggleViewer
+              value={output}
+              environment={environment}
+              {...settings}
+            />
+          )}
         />
       );
     }
