+++ packages/components/src/components/ViewerMenuBar/ViewerBody.tsx
@@ -1,9 +1,8 @@
-import { FC } from "react";
+import { FC, ReactComponentElement } from "react";
 
 import { SqValue } from "@quri/squiggle-lang";
 
 import { SqOutputResult } from "../../../../squiggle-lang/src/public/types.js";
-import { ErrorBoundary } from "../ErrorBoundary.js";
 import { SquiggleErrorAlert } from "../SquiggleErrorAlert.js";
 import { SquiggleViewer } from "../SquiggleViewer/index.js";
 import { ViewerMode } from "./index.js";
@@ -12,6 +11,7 @@ type Props = {
   mode: ViewerMode;
   output: SqOutputResult;
   isRunning: boolean;
+  viewer: (output: SqValue) => ReactComponentElement<typeof SquiggleViewer>;
 };
 
 export function modeToValue(
@@ -34,9 +34,16 @@ export function modeToValue(
     case "AST":
       return;
   }
+  if (mode.tag === "CustomResultPath") {
+    const rootValue =
+      mode.value.root === "result"
+        ? output.value.result
+        : output.value.bindings.asValue();
+    return rootValue.getSubvalueByPath(mode.value, () => undefined);
+  }
 }
 
-export const ViewerBody: FC<Props> = ({ output, mode, isRunning }) => {
+export const ViewerBody: FC<Props> = ({ output, mode, isRunning, viewer }) => {
   if (!output.ok) {
     return <SquiggleErrorAlert error={output.value} />;
   }
@@ -63,10 +70,7 @@ export const ViewerBody: FC<Props> = ({ output, mode, isRunning }) => {
         // `opacity-0 squiggle-semi-appear` would be better, but won't work reliably until we move Squiggle evaluation to Web Workers
         <div className="absolute z-10 inset-0 bg-white opacity-50" />
       )}
-      <ErrorBoundary>
-        {/* we don't pass settings or editor here because they're already configured in `<ViewerProvider>`; hopefully `<SquiggleViewer>` itself won't need to rely on settings, otherwise things might break */}
-        <SquiggleViewer value={usedValue} />
-      </ErrorBoundary>
+      {viewer(usedValue)}
     </div>
   );
 };
